pyproject.toml
[project]
name = "lightning-project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "lightning>=2.5.2",
    "matplotlib>=3.10.3",
    "pyyaml>=6.0.2",
    "tensorboard>=2.20.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
